+++ server/tests/websocket/connection.test.ts
@@ -55,7 +55,7 @@ import { resetTestDb, resetRateLimits } from '../helpers/test-db';
 import { createUser, createTrip } from '../helpers/factories';
 import { authCookie } from '../helpers/auth';
 import { setupWebSocket } from '../../src/websocket';
-import { createEphemeralToken } from '../../src/services/ephemeralTokens';
+import { createEphemeralToken } from '../../src/nest/auth/ephemeral-tokens';
 import { createWsToken } from '../../src/nest/auth/auth.bridge';
 
 let server: http.Server;
